requirements.txt
fastapi
uvicorn>=0.26.0
openai>=0.10.0
langchain>=0.1.9
pydantic>=1.8.0
chromadb>=0.4.0
pandas>=1.3.0
sentence-transformers>=2.0.0
numpy>=1.21.0
retriever>=0.2.0
gcloud
python-docx
google-auth
google-auth-oauthlib
google-api-python-client
dvc
pydrive
tiktoken